# Improve primitive filter performance