--- a/src/types/bit_chunk.rs
+++ b/src/types/bit_chunk.rs
@@ -129,6 +129,12 @@ impl<T: BitChunk> BitChunkOnes<T> {
             remaining: value.count_ones() as usize,
         }
     }
+
+    #[inline]
+    #[cfg(feature = "compute_filter")]
+    pub(crate) fn from_known_count(value: T, remaining: usize) -> Self {
+        Self { value, remaining }
+    }
 }
 
 impl<T: BitChunk> Iterator for BitChunkOnes<T> {